requirements.txt
alembic==1.14.0
blinker==1.9.0
click==8.1.8
Faker==33.3.1
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
setuptools==75.8.0
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
